Renames rails-js_route_dispatch -> rails-js-routes
