+++ lib/rails/js/routes/version.rb
@@ -1,0 +1,7 @@
+module Rails
+  module JS
+    module Routes
+      VERSION = "0.2.0"
+    end
+  end
+end
