--- lib/rails/js_route_dispatch.rb
@@ -1,8 +1,0 @@
-module Rails
-  module JsRouteDispatch
-    if defined?(::Rails)
-      require_relative 'js_route_dispatch/view_helpers'
-      require_relative 'js_route_dispatch/engine'
-    end
-  end
-end
