--- lib/rails/js_route_dispatch/version.rb
@@ -1,5 +1,0 @@
-module Rails
-  module JsRouteDispatch
-    VERSION = "0.1.0"
-  end
-end
