+++ rails-js-routes.gemspec
@@ -1,11 +1,11 @@
 $:.push File.expand_path("../lib", __FILE__)
 
 # Maintain your gem's version:
-require "rails/js_route_dispatch/version"
+require "rails/js/routes/version"
 
 Gem::Specification.new do |s|
-  s.name          = 'rails-js_route_dispatch'
-  s.version       = Rails::JsRouteDispatch::VERSION
+  s.name          = 'rails-js-routes'
+  s.version       = Rails::JS::Routes::VERSION
   s.date          = Time.now.strftime("%Y-%m-%d")
   s.summary       = "A new approach to structure your javascript in non single-page Rails apps"
   s.description   = "Divide your javascript like you divide your ruby code. Put it in controllers and actions and write clean, scalable code."
@@ -13,7 +13,7 @@ Gem::Specification.new do |s|
   s.email         = 'finn@heemeyer.net'
   s.files         = `git ls-files`.split("\n")
   s.require_paths = ['lib']
-  s.homepage      = 'http://rubygems.org/gems/rails-js_route_dispatch'
+  s.homepage      = 'http://rubygems.org/gems/rails-js-routes'
   s.license       = 'MIT'
 
   s.add_dependency "rails", "~> 4"
