+++ test/dummy/app/assets/javascripts/application.js
@@ -10,5 +10,5 @@
 // Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
 // about supported directives.
 //
-//= require route_dispatch
+//= require routes
 //= require_tree .
