+++ test/dummy/app/views/layouts/application.html.erb
@@ -5,7 +5,7 @@
   <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track' => true %>
   <%= javascript_include_tag 'application', 'data-turbolinks-track' => true %>
   <%= csrf_meta_tags %>
-  <%= route_dispatch %>
+  <%= js_routes %>
 </head>
 <body>
